Improve switching ability
Don't keep adding stylesheets but instead re-using existing elements.
If already using a google stylesheet with a custom sheet, preserves
ordering of stylesheets so that custom style overrides remain in effect.

Requirement is that if already using google stylesheet in page, then it
must be using the same version as the themeswitcher is told to use